requirements.txt
duckdb
sqlalchemy
pandas
python-dotenv
requests
beautifulsoup4
pyarrow
typing-extensions
tqdm
openpyxl
lxml
